dropped keymap.h, and changed the KeySym typedef to unsigned int
keymap.h had redundant declarations (KeySym again), and a prototype for
a function used once in linux-specific code, so I decided to drop it,
and fwd-declare the function.

KeySyms are 32bit, no need for unsigned long.

+++ src/kbemu.h
@@ -25,7 +25,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 void kbemu_set_display(Display *dpy);
 #else
 /* When X11 is not available, define KeySym for API compatibility */
-typedef unsigned long KeySym;
+typedef unsigned int KeySym;
 #endif
 
 /* Initialize keyboard emulation based on config */
